+++ src/js/td-scenes/battle/animation/game-state/reflect/lightning/death-lightning.ts
@@ -1,0 +1,26 @@
+import { all } from "../../../../../../animation/all";
+import { Animate } from "../../../../../../animation/animate";
+import { delay } from "../../../../../../animation/delay";
+import { ReflectAnimationParam } from "../animation-param";
+
+/**
+ * 電撃バリア 死亡
+ * @param param アニメーションパラメータ
+ * @returns アニメーション
+ */
+export function deathLightning(param: ReflectAnimationParam): Animate {
+  return all(
+    param.reflecting.hud.resultIndicator
+      .slideIn()
+      .chain(delay(700))
+      .chain(param.reflecting.hud.resultIndicator.moveToEdge()),
+    param.damaged.td.hitMark.lightning.popUp(),
+    delay(100).chain(
+      all(
+        param.damaged.sprite.down(),
+        param.damaged.td.damageIndicator.popUp(param.effect.damage),
+        param.damaged.hud.gauge.hp(param.damaged.state.armdozer.hp),
+      ),
+    ),
+  );
+}
